update storage handler
