+++ bin/daemon.rs
@@ -1,4 +1,4 @@
-use bns_core::message::handler::MessageHandler;
+use bns_core::message::MessageHandler;
 use bns_core::session::SessionManager;
 use bns_core::swarm::Swarm;
 use bns_core::types::message::MessageListener;
