+++ bns-core/src/dht/mod.rs
@@ -4,7 +4,7 @@ pub use did::Did;
 mod chord;
 mod types;
 pub use {chord::PeerRing, chord::PeerRingAction, chord::RemoteAction as PeerRingRemoteAction};
-pub use {types::Chord, types::ChordStablize};
+pub use {types::Chord, types::ChordStablize, types::ChordStorage};
 
 mod stabilization;
 pub use stabilization::Stabilization;
