+++ bns-core/tests/default/test_message_handler.rs
@@ -4,7 +4,7 @@ pub mod test {
     use bns_core::ecc::SecretKey;
     use bns_core::err::Result;
     use bns_core::message;
-    use bns_core::message::handler::MessageHandler;
+    use bns_core::message::MessageHandler;
     use bns_core::message::{Message, MessageRelayMethod};
     use bns_core::session::SessionManager;
     use bns_core::swarm::Swarm;
